osmaxx/pyproject.toml
[tool.poetry]
name = "osmaxx"
version = "5.0.0"
description = ""
authors = ["Geometa Lab <geometalab@ost.ch>"]

[tool.poetry.dependencies]
python = ">=3.10,<4"
"geometalab.osm-pbf-file-size-estimation-service" = "^3.0.0"
drf-extensions = "*"
djangorestframework-gis = "*"
djangorestframework = "*"
django-crispy-forms = "*"
django-debug-toolbar = "*"
django-downloadview = "*"
django-environ = "*"
django-extensions = "*"
django-filter = "*"
django-model-utils = "*"
django-secure = "*"
djangorestframework-jwt = "*"
furl = "*"
GeoAlchemy2 = "*"
Markdown = "*"
memoize = "*"
osmium = "*"
psycopg2-binary = "*"
requests = "*"
sqlalchemy = "<2"
SQLAlchemy-Utils = "*"
sqlalchemy-views = "*"
Werkzeug = "*"
whitenoise = "*"
python-social-auth = {extras = ["django"], version = "*"}
sentry-sdk = "*"
scipy = "*"
numpy = "*"
ipython = "*"
gunicorn = "*"
posthog = "*"
celery = {extras = ["redis", "auth", "msgpack"], version = "*"}
django-celery-results = "*"
django-celery-beat = "*"
pymemcache = "*"
social-auth-app-django = "*"
crispy-bootstrap3 = "^2022.1"

[tool.poetry.dev-dependencies]
pytest-django = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-selenium = "*"
flake8 = "*"
coverage = "*"
mkdocs = "*"
PyHamcrest = "*"
vcrpy = "*"
requests-mock = "*"
Jinja2 = "*"
"ruamel.yaml" = "*"
pypandoc = "*"
black = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
    "integration",
]
DJANGO_SETTINGS_MODULE = ''
python_files = [
    "tests",
    "tests.py",
    "test_*.py",
    "*_tests.py",
]
